requirements.txt
# Audio
sounddevice>=0.4.6
pydub>=0.25.1
audioop-lts>=0.2.2  # Required for pydub on Python 3.13+
numpy>=1.26.0

# API Clients
groq>=0.4.0
openai>=1.0.0

# System Integration
pynput>=1.7.6
pyperclip>=1.8.2
pystray>=0.19.4
Pillow>=10.0.0

# UI
customtkinter>=5.2.0

# Notifications
plyer>=2.1.0

# Config
python-dotenv>=1.0.0

# Dev / Testing
pytest>=7.0.0
